degreedaddy/requirements.txt
Flask
openai
python-docx
fpdf
python-dotenv
nest_asyncio
